Updated calendar_ms
Fully functioning calendar.
Updated the homepage nav bar to link to this page

Notice. Before accessing this page, run the calendarDB.sql script first to set up the DB. testdata.sql is optional.

Main functionalities of the calendar.
Create, Update, Delete, Drag and drop, Select multiple cells at once, Resize (reschedule) with the mouse

Explanation of the code
https://www.youtube.com/watch?v=_lH9ZwAsvWM

TO BE DONE:
Change to AMQP
Tie the calendar to the ID of user, individual calendars. (Need to create a table to link these)
Create the DB that will store the selections of the individuals so that it may be used by this microservice.

TBCs are just things that we might want to consider. Best practices from WAD etc. If not necessary just remove.