Updated homepage.php, Created payment_ms, Created search.php template
Testing the use of search bar in homepage.php. Still figuring out the bootstrap framework. Might want to code it out in a separate folder, so placed it in search_ms folder.

Created the payment_ms to mimic the credit card details form. Will need to relook at the code to understand. Just put for now.

Created the googlemaps_sample.php. Just a sample page to show how the google maps is invoked.